requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.4.0
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy
Werkzeug==3.0.1